Heart Disease Detection using Machine Learning
In the last few decades, cardiovascular diseases have exceeded all others as the main cause of death worldwide, in both developed and developing countries. With early diagnosis and ongoing clinical monitoring, the mortality rate due to heart diseases can be reduced. In this study, we investigated four machine learning techniques for detecting cardiac problems. This project's primary objective is to provide clinicians with a diagnostic tool for early detection of heart issues. So, it will be easier to effectively treat patients and keep them from getting hurt badly.